refactor: 자동차 이름 적합한지 검사하는 기능 클래스 분리
이름 길이 제한, 중복검사 기능 RacingGameRule클래스로 분리
